+++ mrfmsim/component/grid.py
@@ -28,23 +28,23 @@ class Grid(ComponentBase):
         based on the rounded value of grid shape and step size.
     """
 
-    shape: tuple[int]
-    step: list[float] = field(metadata={"unit": "nm"})
-    origin: list[float] = field(metadata={"unit": "nm"})
-    voxel: float = field(init=False, metadata={"unit": "nm^3"})
-    range: np.array = field(init=False, metadata={"unit": "nm"})
-    length: np.array = field(init=False, metadata={"unit": "nm"})
+    grid_shape: tuple[int]
+    grid_step: list[float] = field(metadata={"unit": "nm", "format": ".1f"})
+    grid_origin: list[float] = field(metadata={"unit": "nm", "format": ".1f"})
+    grid_voxel: float = field(init=False, metadata={"unit": "nm^3"})
+    grid_range: np.array = field(init=False, metadata={"unit": "nm", "format": ".1f"})
+    grid_length: np.array = field(init=False, metadata={"unit": "nm", "format": ".1f"})
 
     def __post_init__(self):
         """Calculate grid parameters."""
 
-        self.voxel = np.array(self.step).prod()
-        self.range = (np.array(self.shape) - [1, 1, 1]) * self.step
-        self.length = np.array(self.shape) * np.array(self.step)
-        self.extents = self.grid_extents(self.range, self.origin)
+        self.grid_voxel = np.array(self.grid_step).prod()
+        self.grid_range = (np.array(self.grid_shape) - [1, 1, 1]) * self.grid_step
+        self.grid_length = np.array(self.grid_shape) * np.array(self.grid_step)
+        self.grid_extents = self.grid_extents_method(self.grid_range, self.grid_origin)
 
     @staticmethod
-    def grid_extents(length, origin):
+    def grid_extents_method(length, origin):
         """Calculate grid extents based on the grid length and origin.
 
         The result is column stacked into a dimension of (3, 2)
@@ -59,12 +59,12 @@ class Grid(ComponentBase):
         The benefit of the property is that it generates the grid array at run time.
         """
 
-        # extents = self.grid_extents(self.range, self.origin)
+        # extents = self.grid_extents(self.grid_range, self.origin)
 
         return np.ogrid[
-            self.extents[0][0] : self.extents[0][1] : self.shape[0] * 1j,
-            self.extents[1][0] : self.extents[1][1] : self.shape[1] * 1j,
-            self.extents[2][0] : self.extents[2][1] : self.shape[2] * 1j,
+            self.grid_extents[0][0] : self.grid_extents[0][1] : self.grid_shape[0] * 1j,
+            self.grid_extents[1][0] : self.grid_extents[1][1] : self.grid_shape[1] * 1j,
+            self.grid_extents[2][0] : self.grid_extents[2][1] : self.grid_shape[2] * 1j,
         ]
 
     def extend_grid_by_points(self, ext_pts):
@@ -75,9 +75,9 @@ class Grid(ComponentBase):
             three dimensions.
         """
 
-        ext_shape = self.shape + np.array(ext_pts) * 2
-        ext_range = (ext_shape - [1, 1, 1]) * self.step
-        extents = self.grid_extents(ext_range, self.origin)
+        ext_shape = self.grid_shape + np.array(ext_pts) * 2
+        ext_range = (ext_shape - [1, 1, 1]) * self.grid_step
+        extents = self.grid_extents_method(ext_range, self.grid_origin)
 
         return np.ogrid[
             extents[0][0] : extents[0][1] : ext_shape[0] * 1j,
@@ -96,5 +96,5 @@ class Grid(ComponentBase):
             three dimensions.
         """
 
-        pts = np.floor(np.array(ext_length) / self.step).astype(int)
+        pts = np.floor(np.array(ext_length) / self.grid_step).astype(int)
         return self.extend_grid_by_points(pts)